Add the Excluded property to the TunnelConnectRequest event
